map scheint mit dem iterator ein problem zu haben. find findet korrekte werte, for-schleife gibt 1.0e+6 aus

+++ src/main/DeviceVerwalter.h
@@ -68,7 +68,7 @@ class DeviceVerwalter {
   /**
  * checks, if ingredient is double
  */
-  int checkForDouble(std::string);
+  float checkForDouble(std::string);
   /**
 * test for special ingredient
 */
